+++ src/app/admin/_components/view/add.tsx
@@ -1,5 +1,7 @@
+"use client";
+
 import Head from "next/head";
-import { useRouter } from "next/router";
+import { useRouter } from "next/navigation";
 import { FormEvent, useState } from "react";
 import { AdminLayout } from "@/features/admin/components/AdminLayout";
 import { bookRepository } from "@/infrastructure/book";
