+++ src/app/admin/add/page.tsx
@@ -1,0 +1,9 @@
+import Add from "../_components/view/add";
+
+export default function Page() {
+  return (
+    <>
+      <Add />
+    </>
+  );
+}
