+++ src/app/admin/edit/page.tsx
@@ -1,0 +1,5 @@
+import Edit from "../_components/view/edit";
+
+export default function Page() {
+  return <Edit />;
+}
