+++ src/app/admin/page.tsx
@@ -1,0 +1,5 @@
+import Admin from "./_components/view";
+
+export default function Page() {
+  return <Admin />;
+}
