+++ src/app/login/page.tsx
@@ -1,0 +1,9 @@
+import Login from "./_components/view";
+
+export default function Page() {
+  return (
+    <>
+      <Login />
+    </>
+  );
+}
