+++ src/app/signup/completed/page.tsx
@@ -1,0 +1,9 @@
+import Completed from "../_components/view/completed";
+
+export default function Page() {
+  return (
+    <>
+      <Completed />
+    </>
+  );
+}
