+++ src/app/signup/page.tsx
@@ -1,0 +1,9 @@
+import SignUp from "./_components/view";
+
+export default function Page() {
+  return (
+    <>
+      <SignUp />
+    </>
+  );
+}
